Restructure cli app into modules
